requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1
redis>=5.0.4
cryptography>=43.0.0
orjson>=3.10.7
httpx>=0.27.2
pytest>=8.2.0
pytest-cov>=5.0.0
ruff>=0.6.8
click>=8.1.7
rich>=13.7.1
hypothesis>=6.104.1
prometheus_client>=0.21.0
grpcio>=1.66.0
grpcio-tools>=1.66.0
cyclonedx-bom>=4.5.0
pre-commit>=3.7.0
PyYAML>=6.0.2
